@@ -1,4 +1,12 @@
-import { CreationOptional, DataTypes, ForeignKey, InferAttributes, InferCreationAttributes, Model, NonAttribute } from "sequelize";
+import {
+  CreationOptional,
+  DataTypes,
+  ForeignKey,
+  InferAttributes,
+  InferCreationAttributes,
+  Model,
+  NonAttribute,
+} from "sequelize";
 import User from "./user";
 import db from "../db";
 
@@ -31,10 +39,13 @@ import db from "../db";
  * @property {UUID} [UserId]
  * @memberof module:Models
  */
-class Feedback extends Model<InferAttributes<Feedback>, InferCreationAttributes<Feedback>> {
+class Feedback extends Model<
+  InferAttributes<Feedback>,
+  InferCreationAttributes<Feedback>
+> {
   declare id: CreationOptional<string>;
   declare stars: number;
-  declare text?: string;
+  declare text: string;
 
   declare createdAt: CreationOptional<Date>;
   declare updatedAt: CreationOptional<Date>;
@@ -70,7 +81,7 @@ Feedback.init(
   {
     sequelize: db,
     paranoid: true,
-  }
+  },
 );
 
 export default Feedback;
