@@ -8,7 +8,12 @@ import { NextFunction, Request, Response } from "express";
  * @param {Response} res Outgoing response object
  * @param {Function} next Next handler function (not called)
  */
-function errorHandlingMiddleWare(error: Error, req: Request, res: Response, next: NextFunction) {
+function errorHandlingMiddleWare(
+  error: Error,
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) {
   if (!res.headersSent) {
     res.status(500).render("../src/views/error.ejs", {
       action: "generic error handling",
