@@ -41,16 +41,20 @@ const router = Router();
  *       401:
  *         $ref: '#/components/responses/UnauthorizedError'
  */
-router.post("/", AuthService.authenticateUser(false), (req: Request, res: Response, next: NextFunction) => {
-  const { stars, text } = req.body;
-  const UserId = req.UserId;
-  FeedbackService.create(stars, text, UserId)
-    .then((feedback: Feedback) => {
-      res.send(feedback);
-      return next();
-    })
-    .catch((e: Error) => next(e));
-});
+router.post(
+  "/",
+  AuthService.authenticateUser(false),
+  (req: Request, res: Response, next: NextFunction) => {
+    const { stars, text } = req.body;
+    const UserId = req.UserId;
+    FeedbackService.create(stars, text, UserId)
+      .then((feedback: Feedback) => {
+        res.send(feedback);
+        return next();
+      })
+      .catch((e: Error) => next(e));
+  },
+);
 
 /**
  * @openapi
@@ -73,19 +77,23 @@ router.post("/", AuthService.authenticateUser(false), (req: Request, res: Respon
  *       401:
  *         $ref: '#/components/responses/UnauthorizedError'
  */
-router.get("/", AuthService.authenticateUser(false), (req: Request, res: Response, next: NextFunction) => {
-  const UserId = req.UserId;
-  if (!UserId) {
-    res.send([]);
-    return next();
-  } else {
-    FeedbackService.getAll(UserId)
-      .then((feedbackList: Feedback[]) => {
-        res.send(feedbackList);
-        return next();
-      })
-      .catch((e: Error) => next(e));
-  }
-});
+router.get(
+  "/",
+  AuthService.authenticateUser(false),
+  (req: Request, res: Response, next: NextFunction) => {
+    const UserId = req.UserId;
+    if (!UserId) {
+      res.send([]);
+      return next();
+    } else {
+      FeedbackService.getAll(UserId)
+        .then((feedbackList: Feedback[]) => {
+          res.send(feedbackList);
+          return next();
+        })
+        .catch((e: Error) => next(e));
+    }
+  },
+);
 
 export { router as feedbackRouter };
