@@ -27,14 +27,18 @@ const router = Router();
  *       401:
  *         $ref: '#/components/responses/UnauthorizedError'
  */
-router.get("/", AuthService.authenticateUser(), (req: Request, res: Response, next: NextFunction) => {
-  return SeasonService.getAll(req.UserId)
-    .then((seasonList: Season[]) => {
-      res.send(seasonList);
-      return next();
-    })
-    .catch((e: Error) => next(e));
-});
+router.get(
+  "/",
+  AuthService.authenticateUser(),
+  (req: Request, res: Response, next: NextFunction) => {
+    return SeasonService.getAll(req.UserId)
+      .then((seasonList: Season[]) => {
+        res.send(seasonList);
+        return next();
+      })
+      .catch((e: Error) => next(e));
+  },
+);
 
 /**
  * @openapi
@@ -69,14 +73,18 @@ router.get("/", AuthService.authenticateUser(), (req: Request, res: Response, ne
  *       401:
  *         $ref: '#/components/responses/UnauthorizedError'
  */
-router.post("/", AuthService.authenticateUser(), (req: Request, res: Response, next: NextFunction) => {
-  const { name, year } = req.body;
-  return SeasonService.create(name, year, req.UserId)
-    .then((season: Season) => {
-      res.send(season);
-      return next();
-    })
-    .catch((e: Error) => next(e));
-});
+router.post(
+  "/",
+  AuthService.authenticateUser(),
+  (req: Request, res: Response, next: NextFunction) => {
+    const { name, year } = req.body;
+    return SeasonService.create(name, year, req.UserId)
+      .then((season: Season) => {
+        res.send(season);
+        return next();
+      })
+      .catch((e: Error) => next(e));
+  },
+);
 
 export { router as seasonRouter };
