@@ -19,7 +19,10 @@ class AdminService {
    */
   async findOrCreate(username: string, password: string) {
     logger.debug(
-      `AdminService findOrCreate ${JSON.stringify({ username, password: password ? "<redacted>" : "undefined" })}`
+      `AdminService findOrCreate ${JSON.stringify({
+        username,
+        password: password ? "<redacted>" : "undefined",
+      })}`,
     );
     const [admin, _created] = await Admin.findOrCreate({
       where: { username },
@@ -38,8 +41,10 @@ class AdminService {
    * @param {string} [options.scope="defaultScope"] - Sequelize scope to use.
    * @returns {Promise<Admin|null>} The found Admin or null if not found.
    */
-  findByUsername(username: string, { scope = "defaultScope" }) {
-    logger.debug(`AdminService findByUsername ${JSON.stringify({ username, scope })}`)
+  findByUsername(username: string, { scope = "defaultScope" } = {}) {
+    logger.debug(
+      `AdminService findByUsername ${JSON.stringify({ username, scope })}`,
+    );
     return Admin.scope(scope).findOne({ where: { username } }); // njsscan-ignore: node_nosqli_injection
   }
 
@@ -49,7 +54,7 @@ class AdminService {
    * @returns {Promise<Admin|null>} The found Admin or null if not found.
    */
   findById(id: string) {
-    logger.debug(`AdminService findById ${JSON.stringify({ id })}`)
+    logger.debug(`AdminService findById ${JSON.stringify({ id })}`);
     return Admin.findByPk(id);
   }
 
@@ -59,7 +64,7 @@ class AdminService {
    * @returns {Promise<number>} The count of Admins.
    */
   getCount() {
-    logger.debug(`AdminService getCount`)
+    logger.debug(`AdminService getCount`);
     return Admin.count();
   }
 
@@ -69,7 +74,7 @@ class AdminService {
    * @returns {Promise<Admin[]>} Array of all Admins.
    */
   getAll() {
-    logger.debug(`AdminService getAll`)
+    logger.debug(`AdminService getAll`);
     return Admin.findAll();
   }
 
@@ -79,16 +84,20 @@ class AdminService {
    * @returns {Promise<number>} The net number of Admins added in the last 30 days.
    */
   getTrend() {
-    logger.debug(`AdminService getTrend`)
+    logger.debug(`AdminService getTrend`);
     return Promise.all([
       Admin.count({
         where: {
-          createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          createdAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
       Admin.count({
         where: {
-          deletedAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          deletedAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
     ]).then(([created, deleted]) => created - deleted);
@@ -103,8 +112,13 @@ class AdminService {
    * @throws {Error} If no Admin with the given ID is found.
    */
   update(id: string, data: object & Record<string, unknown>) {
-    const { password, ...logdata } = data
-    logger.debug(`AdminService update ${JSON.stringify({ id, data: { ...logdata, password: password ? "<redacted>" : "undefined" } })}`)
+    const { password, ...logdata } = data;
+    logger.debug(
+      `AdminService update ${JSON.stringify({
+        id,
+        data: { ...logdata, password: password ? "<redacted>" : "undefined" },
+      })}`,
+    );
     return this.findById(id).then(async (admin: Admin | null) => {
       if (admin) {
         logger.debug(`AdminService update ${JSON.stringify({ id, data })}`);
@@ -112,12 +126,8 @@ class AdminService {
         await admin.save();
         return this.findById(id);
       } else {
-        logger.error(
-          `No admin found with ID ${id} when updating`
-        );
-        throw new Error(
-          `No admin found with ID ${id} when updating`
-        );
+        logger.error(`No admin found with ID ${id} when updating`);
+        throw new Error(`No admin found with ID ${id} when updating`);
       }
     });
   }
@@ -130,18 +140,14 @@ class AdminService {
    * @throws {Error} If no Admin with the given ID is found.
    */
   remove(id: string) {
-    logger.debug(`AdminService remove ${JSON.stringify({ id })}`)
+    logger.debug(`AdminService remove ${JSON.stringify({ id })}`);
     return this.findById(id).then(async (admin: Admin | null) => {
       if (admin) {
         logger.debug(`AdminService remove ${JSON.stringify({ id })}`);
         return admin.destroy();
       } else {
-        logger.error(
-          `No admin found with ID ${id} when deleting`
-        );
-        throw new Error(
-          `No admin found with ID ${id} when deleting`
-        );
+        logger.error(`No admin found with ID ${id} when deleting`);
+        throw new Error(`No admin found with ID ${id} when deleting`);
       }
     });
   }
