@@ -3,14 +3,14 @@ import Club from "../db/models/club";
 import Season from "../db/models/season";
 import SeasonTeam from "../db/models/seasonTeam";
 import Team from "../db/models/team";
+import ChoreoService from "./ChoreoService";
+import HitService from "./HitService";
+import MemberService from "./MemberService";
+import SeasonService from "./SeasonService";
+import TeamService from "./TeamService";
 
 const { Op } = require("sequelize");
 const { logger } = require("../plugins/winston");
-const ChoreoService = require("./ChoreoService");
-const HitService = require("./HitService");
-const MemberService = require("./MemberService");
-const SeasonService = require("./SeasonService");
-const TeamService = require("./TeamService");
 
 const defaultInclude = [
   {
@@ -47,9 +47,9 @@ class ClubService {
    * @returns {Club[]}
    */
   async getAll(UserId: string | null, options = { all: false }) {
-    logger.debug(`ClubService getAll ${JSON.stringify({ UserId, options })}`)
+    logger.debug(`ClubService getAll ${JSON.stringify({ UserId, options })}`);
     return Club.findAll({
-      where: options.all ? {} : (UserId ? { UserId } : {}),
+      where: options.all ? {} : UserId ? { UserId } : {},
       include: defaultInclude,
       order: [
         ["createdAt", "ASC"],
@@ -77,7 +77,7 @@ class ClubService {
    * @returns {number}
    */
   getCount() {
-    logger.debug(`ClubService getCount`)
+    logger.debug(`ClubService getCount`);
     return Club.count();
   }
 
@@ -87,16 +87,20 @@ class ClubService {
    * @returns {number}
    */
   getTrend() {
-    logger.debug(`ClubService getTrend`)
+    logger.debug(`ClubService getTrend`);
     return Promise.all([
       Club.count({
         where: {
-          createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          createdAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
       Club.count({
         where: {
-          deletedAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          deletedAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
     ]).then(([created, deleted]) => created - deleted);
@@ -111,7 +115,7 @@ class ClubService {
    * @returns {Club}
    */
   async findById(id: string, UserId: string) {
-    logger.debug(`ClubService findById ${JSON.stringify({ id, UserId })}`)
+    logger.debug(`ClubService findById ${JSON.stringify({ id, UserId })}`);
     return Club.findOne({
       where: { id, UserId },
       include: defaultInclude,
@@ -143,7 +147,7 @@ class ClubService {
    * @returns {Club}
    */
   async findByName(name: string, UserId: string) {
-    logger.debug(`ClubService findByName ${JSON.stringify({ name, UserId })}`)
+    logger.debug(`ClubService findByName ${JSON.stringify({ name, UserId })}`);
     return Club.findAll({
       where: { name, UserId },
       include: defaultInclude,
@@ -172,7 +176,7 @@ class ClubService {
           club = foundClub;
         }
         return club;
-      })
+      }),
     );
   }
 
@@ -185,65 +189,69 @@ class ClubService {
    * @returns {void}
    */
   async seedDemo(club: Club, UserId: string) {
-    logger.debug(`ClubService seedDemo ${JSON.stringify({ club, UserId })}`)
-    SeasonService.getAll(null).then((seasons: Season[]) => {
+    logger.debug(`ClubService seedDemo ${JSON.stringify({ club, UserId })}`);
+    return SeasonService.getAll(null).then((seasons: Season[]) => {
       const currentSeason = seasons.find(
-        (s) => s.year == new Date().getFullYear()
+        (s) => s.year == new Date().getFullYear(),
       ) as Season;
-      TeamService.create("Demo-Team", club.id, currentSeason.id, UserId).then(
-        (team: Team) => {
-          const seasonTeam = team.SeasonTeams[0];
-          Promise.all([
-            MemberService.create(
-              "Tina Turnerin",
-              "Tini",
-              "T",
-              seasonTeam.id,
-              UserId
-            ),
-            MemberService.create(
-              "Zoe Zuverlässig",
-              "Zoe",
-              "Z",
-              seasonTeam.id,
-              UserId
-            ),
-            MemberService.create(
-              "Fenja Flyer",
-              "Fipsi",
-              "F",
-              seasonTeam.id,
-              UserId
-            ),
-          ]).then((members) =>
-            ChoreoService.create(
-              "Demo-Choreo",
-              25,
-              undefined,
-              seasonTeam.id,
-              members.map((m) => ({ id: m.id })),
-              UserId
-            ).then((choreo: Choreo) => {
-              Promise.all([
-                HitService.create(
-                  "Pose",
-                  0,
-                  choreo.id,
-                  members.map((m) => m.id),
-                  UserId
-                ),
-                HitService.create(
-                  "Flick-Flack",
-                  2,
-                  choreo.id,
-                  [members[0].id],
-                  UserId
-                ),
-              ]);
-            })
-          );
-        }
-      );
+      return TeamService.create(
+        "Demo-Team",
+        club.id,
+        currentSeason.id,
+        UserId,
+      ).then((team: Team | null) => {
+        if (!team) return;
+        const seasonTeam = team.SeasonTeams[0];
+        Promise.all([
+          MemberService.create(
+            "Tina Turnerin",
+            "Tini",
+            "T",
+            seasonTeam.id,
+            UserId,
+          ),
+          MemberService.create(
+            "Zoe Zuverlässig",
+            "Zoe",
+            "Z",
+            seasonTeam.id,
+            UserId,
+          ),
+          MemberService.create(
+            "Fenja Flyer",
+            "Fipsi",
+            "F",
+            seasonTeam.id,
+            UserId,
+          ),
+        ]).then((members) =>
+          ChoreoService.create(
+            "Demo-Choreo",
+            25,
+            undefined,
+            seasonTeam.id,
+            members.map((m) => ({ id: m.id })),
+            UserId,
+          ).then((choreo: Choreo) => {
+            Promise.all([
+              HitService.create(
+                "Pose",
+                0,
+                choreo.id,
+                members.map((m) => m.id),
+                UserId,
+              ),
+              HitService.create(
+                "Flick-Flack",
+                2,
+                choreo.id,
+                [members[0].id],
+                UserId,
+              ),
+            ]);
+          }),
+        );
+      });
     });
   }
 
@@ -257,7 +265,7 @@ class ClubService {
    */
   async findOrCreate(name: string, UserId: string) {
     logger.debug(
-      `ClubService findOrCreate ${JSON.stringify({ name, UserId })}`
+      `ClubService findOrCreate ${JSON.stringify({ name, UserId })}`,
     );
     const [club, _created] = await Club.findOrCreate({
       where: { name, UserId },
@@ -276,9 +284,18 @@ class ClubService {
    * @param {boolean} [options.all=false]
    * @returns {Club}
    */
-  async update(id: string, data: object, UserId: string | null, options = { all: false }) {
-    logger.debug(`ClubService update ${JSON.stringify({ id, data, UserId, options })}`);
-    return Club.findOne({ where: options.all || !UserId ? { id } : { id, UserId } }) // njsscan-ignore: node_nosqli_injection
+  async update(
+    id: string,
+    data: object,
+    UserId: string | null,
+    options = { all: false },
+  ) {
+    logger.debug(
+      `ClubService update ${JSON.stringify({ id, data, UserId, options })}`,
+    );
+    return Club.findOne({
+      where: options.all || !UserId ? { id } : { id, UserId },
+    }) // njsscan-ignore: node_nosqli_injection
       .then(async (foundClub) => {
         if (foundClub) {
           await foundClub.update(data);
@@ -301,8 +318,12 @@ class ClubService {
    * @returns {Promise<void>}
    */
   async remove(id: string, UserId: string | null, options = { all: false }) {
-    logger.debug(`ClubService remove ${JSON.stringify({ id, UserId, options })}`);
-    return Club.findOne({ where: options.all || !UserId ? { id } : { id, UserId } }) // njsscan-ignore: node_nosqli_injection
+    logger.debug(
+      `ClubService remove ${JSON.stringify({ id, UserId, options })}`,
+    );
+    return Club.findOne({
+      where: options.all || !UserId ? { id } : { id, UserId },
+    }) // njsscan-ignore: node_nosqli_injection
       .then((foundClub) => {
         if (foundClub) {
           return foundClub.destroy();
