@@ -1,9 +1,9 @@
 import Feedback from "../db/models/feedback";
 import logger from "../plugins/winston";
+import MailService from "./MailService";
+import UserService from "./UserService";
 
 const { Sequelize, Op } = require("sequelize");
-const MailService = require("./MailService");
-const UserService = require("./UserService");
 
 /**
  * Service for handling user feedback.
@@ -21,16 +21,21 @@ class FeedbackService {
    * @returns {Promise<Feedback>} The created feedback object.
    */
   create(stars: number, text: string, UserId: string) {
-    logger.debug(`FeedbackService create ${JSON.stringify({ stars, text, UserId })}`)
+    logger.debug(
+      `FeedbackService create ${JSON.stringify({ stars, text, UserId })}`,
+    );
     return Feedback.create({ stars, text, UserId }).then(async (feedback) => {
       let user = null;
       if (UserId) user = await UserService.findById(UserId).catch(() => null);
-      MailService.sendFeedbackNotice(
-        user?.username,
-        user?.email,
-        feedback.stars,
-        feedback.text
-      );
+      if (!user)
+        throw new Error("User does not exist in order to create this feedback");
+      if (user.email)
+        MailService.sendFeedbackNotice(
+          user?.username,
+          user?.email,
+          feedback.stars,
+          feedback.text,
+        );
       return feedback;
     });
   }
@@ -42,7 +47,7 @@ class FeedbackService {
    * @returns {Promise<Feedback[]>} List of feedbacks for the user.
    */
   getAll(UserId: string) {
-    logger.debug(`FeedbackService getAll ${JSON.stringify({ UserId })}`)
+    logger.debug(`FeedbackService getAll ${JSON.stringify({ UserId })}`);
     return Feedback.findAll({ where: { UserId } });
   }
 
@@ -52,7 +57,7 @@ class FeedbackService {
    * @returns {Promise<Feedback>} The most recently created feedback.
    */
   getNewest() {
-    logger.debug(`FeedbackService getNewest`)
+    logger.debug(`FeedbackService getNewest`);
     return Feedback.findAll({ order: ["createdAt"] }).then((feedbackList) => {
       return feedbackList[0];
     });
@@ -64,7 +69,7 @@ class FeedbackService {
    * @returns {Promise<number>} The average rating of all feedbacks.
    */
   getTotalAverage() {
-    logger.debug(`FeedbackService getTotalAverage`)
+    logger.debug(`FeedbackService getTotalAverage`);
     return Feedback.findAll({
       attributes: [[Sequelize.fn("avg", Sequelize.col("stars")), "stars"]],
     }).then((result) => parseFloat(String(result[0].dataValues.stars)));
@@ -76,7 +81,7 @@ class FeedbackService {
    * @returns {Promise<number>} The average rating of feedbacks from the last month.
    */
   getAverageOfLastMonth() {
-    logger.debug(`FeedbackService getAverageOfLastMonth`)
+    logger.debug(`FeedbackService getAverageOfLastMonth`);
     return Feedback.findAll({
       where: {
         createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
