@@ -1,3 +1,4 @@
+import logger from "@/plugins/winston";
 import { sendMail, verify } from "../plugins/nodemailer";
 const { timeStringToMillis } = require("../utils/time");
 const i18n = require("i18n");
@@ -29,8 +30,18 @@ class MailService {
    * @param {string} userEmail - The email of the new user.
    * @returns {Promise<void[]>} Resolves when all emails are sent.
    */
-  sendUserRegistrationNotice(username: string, userId: string, userEmail: string) {
-    logger.debug(`MailService sendUserRegistrationNotice ${JSON.stringify({ username, userId, userEmail })}`)
+  sendUserRegistrationNotice(
+    username: string,
+    userId: string,
+    userEmail: string,
+  ) {
+    logger.debug(
+      `MailService sendUserRegistrationNotice ${JSON.stringify({
+        username,
+        userId,
+        userEmail,
+      })}`,
+    );
     return Promise.all(
       this.adminEmails.map((recipient) => {
         return sendMail(
@@ -49,9 +60,9 @@ class MailService {
               cid: "choreo-planer-icon",
             },
           ],
-          "de"
+          "de",
         );
-      })
+      }),
     );
   }
 
@@ -63,8 +74,20 @@ class MailService {
    * @param {string} locale - The locale for the email content.
    * @returns {Promise<void>} Resolves when the email is sent.
    */
-  sendWelcomeEmail(username: string, userId: string, userEmail: string, locale: string) {
-    logger.debug(`MailService sendWelcomeEmail ${JSON.stringify({ username, userId, userEmail, locale })}`)
+  sendWelcomeEmail(
+    username: string,
+    userId: string,
+    userEmail: string,
+    locale: string,
+  ) {
+    logger.debug(
+      `MailService sendWelcomeEmail ${JSON.stringify({
+        username,
+        userId,
+        userEmail,
+        locale,
+      })}`,
+    );
     return sendMail(
       userEmail,
       i18n.__({ phrase: "mail.welcome.willkommen-beim-choreo-planer", locale }),
@@ -81,7 +104,7 @@ class MailService {
           cid: "choreo-planer-icon",
         },
       ],
-      locale
+      locale,
     );
   }
 
@@ -93,8 +116,20 @@ class MailService {
    * @param {string} locale - The locale for the email content.
    * @returns {Promise<void>} Resolves when the email is sent.
    */
-  sendEmailConfirmationEmail(username: string, userId: string, userEmail: string, locale: string) {
-    logger.debug(`MailService sendEmailConfirmationEmail ${JSON.stringify({ username, userId, userEmail, locale })}`)
+  sendEmailConfirmationEmail(
+    username: string,
+    userId: string,
+    userEmail: string,
+    locale: string,
+  ) {
+    logger.debug(
+      `MailService sendEmailConfirmationEmail ${JSON.stringify({
+        username,
+        userId,
+        userEmail,
+        locale,
+      })}`,
+    );
     return sendMail(
       userEmail,
       i18n.__({
@@ -115,7 +150,7 @@ class MailService {
           cid: "choreo-planer-icon",
         },
       ],
-      locale
+      locale,
     );
   }
 
@@ -127,8 +162,20 @@ class MailService {
    * @param {string} text - The feedback text provided by the user.
    * @returns {Promise<void[]>} Resolves when all emails are sent.
    */
-  sendFeedbackNotice(username: string, userEmail: string, stars: number, text: string) {
-    logger.debug(`MailService sendFeebackNotice ${JSON.stringify({ username, userEmail, stars, text })}`)
+  sendFeedbackNotice(
+    username: string,
+    userEmail: string,
+    stars: number,
+    text: string,
+  ) {
+    logger.debug(
+      `MailService sendFeebackNotice ${JSON.stringify({
+        username,
+        userEmail,
+        stars,
+        text,
+      })}`,
+    );
     return Promise.all(
       this.adminEmails.map((recipient) => {
         return sendMail(
@@ -148,9 +195,9 @@ class MailService {
               cid: "choreo-planer-icon",
             },
           ],
-          "en"
+          "en",
         );
-      })
+      }),
     );
   }
 
@@ -162,8 +209,20 @@ class MailService {
    * @param {string} locale - The locale for the email content.
    * @returns {Promise<void>} Resolves when the email is sent.
    */
-  sendSsoEmail(userEmail: string, username: string, ssoToken: string, locale: string) {
-    logger.debug(`MailService sendSsoEmail ${JSON.stringify({ userEmail, username, ssoToken, locale })}`)
+  sendSsoEmail(
+    userEmail: string,
+    username: string,
+    ssoToken: string,
+    locale: string,
+  ) {
+    logger.debug(
+      `MailService sendSsoEmail ${JSON.stringify({
+        userEmail,
+        username,
+        ssoToken,
+        locale,
+      })}`,
+    );
     return sendMail(
       userEmail,
       i18n.__({ phrase: "mail.ssoLogin.dein-login-link", locale }),
@@ -174,7 +233,7 @@ class MailService {
         frontendDomain: process.env.FRONTEND_DOMAIN,
         expirationDate: new Date(
           new Date().valueOf() +
-          timeStringToMillis(process.env.SSO_TOKEN_EXPIRES_IN)
+            timeStringToMillis(process.env.SSO_TOKEN_EXPIRES_IN),
         ).toLocaleString(locale, {
           weekday: "long",
           year: "numeric",
@@ -191,7 +250,7 @@ class MailService {
           cid: "choreo-planer-icon",
         },
       ],
-      locale
+      locale,
     );
   }
 
@@ -209,9 +268,17 @@ class MailService {
     username: string,
     ssoToken: string,
     notificationTitle: string | null = null,
-    locale: string
+    locale: string,
   ) {
-    logger.debug(`MailService sendNotificationNotice ${JSON.stringify({ userEmail, username, ssoToken, notificationTitle, locale })}`)
+    logger.debug(
+      `MailService sendNotificationNotice ${JSON.stringify({
+        userEmail,
+        username,
+        ssoToken,
+        notificationTitle,
+        locale,
+      })}`,
+    );
     return sendMail(
       userEmail,
       i18n.__({ phrase: "mail.newNotification.neue-benachrichtigung", locale }),
@@ -223,7 +290,7 @@ class MailService {
         frontendDomain: process.env.FRONTEND_DOMAIN,
         expirationDate: new Date(
           new Date().valueOf() +
-          timeStringToMillis(process.env.SSO_TOKEN_EXPIRES_IN)
+            timeStringToMillis(process.env.SSO_TOKEN_EXPIRES_IN),
         ).toLocaleString(locale, {
           weekday: "long",
           year: "numeric",
@@ -240,7 +307,7 @@ class MailService {
           cid: "choreo-planer-icon",
         },
       ],
-      locale
+      locale,
     );
   }
 }
