@@ -18,7 +18,7 @@ class MemberService {
    * @returns {Promise<Array>} List of members.
    */
   async getAll(UserId: string, options = { all: false }) {
-    logger.debug(`MemberService getAll ${JSON.stringify({ UserId, options })}`)
+    logger.debug(`MemberService getAll ${JSON.stringify({ UserId, options })}`);
     return Member.findAll({ where: options.all ? {} : { UserId } });
   }
 
@@ -29,7 +29,7 @@ class MemberService {
    * @returns {Promise<Object|null>} The member object or null if not found.
    */
   async findById(id: string, UserId: string) {
-    logger.debug(`MemberService findById ${JSON.stringify({ id, UserId })}`)
+    logger.debug(`MemberService findById ${JSON.stringify({ id, UserId })}`);
     return Member.findOne({ where: { id, UserId } }); // njsscan-ignore: node_nosqli_injection
   }
 
@@ -38,7 +38,7 @@ class MemberService {
    * @returns {Promise<number>} The count of members.
    */
   getCount() {
-    logger.debug(`MemberService getCount`)
+    logger.debug(`MemberService getCount`);
     return Member.count();
   }
 
@@ -47,16 +47,20 @@ class MemberService {
    * @returns {Promise<number>} The difference between created and deleted members in the last 30 days.
    */
   getTrend() {
-    logger.debug(`MemberService getTrend`)
+    logger.debug(`MemberService getTrend`);
     return Promise.all([
       Member.count({
         where: {
-          createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          createdAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
       Member.count({
         where: {
-          deletedAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          deletedAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
     ]).then(([created, deleted]) => created - deleted);
@@ -71,7 +75,13 @@ class MemberService {
    * @param {string} UserId - The user ID.
    * @returns {Promise<Object>} The created member object.
    */
-  async create(name: string, nickname: string, abbreviation: string, SeasonTeamId: string, UserId: string) {
+  async create(
+    name: string,
+    nickname: string,
+    abbreviation: string | null,
+    SeasonTeamId: string,
+    UserId: string,
+  ) {
     if (!abbreviation)
       abbreviation = name
         .split(" ")
@@ -85,7 +95,7 @@ class MemberService {
         abbreviation,
         SeasonTeamId,
         UserId,
-      })}`
+      })}`,
     );
     return Member.create({
       name,
@@ -105,7 +115,13 @@ class MemberService {
    * @param {string} UserId - The user ID.
    * @returns {Promise<Object>} The found or created member object.
    */
-  async findOrCreate(name: string, nickname: string, abbreviation: string, SeasonTeamId: string, UserId: string) {
+  async findOrCreate(
+    name: string,
+    nickname: string | null,
+    abbreviation: string | null,
+    SeasonTeamId: string,
+    UserId: string,
+  ) {
     logger.debug(
       `MemberService findOrCreate ${JSON.stringify({
         name,
@@ -113,7 +129,7 @@ class MemberService {
         abbreviation,
         SeasonTeamId,
         UserId,
-      })}`
+      })}`,
     );
 
     const defaultAbbreviation = name
@@ -121,14 +137,23 @@ class MemberService {
       .map((s) => s.substring(0, 1))
       .join("");
 
+    const whereClause: {
+      name: string;
+      abbreviation: string;
+      SeasonTeamId: string;
+      UserId: string;
+      nickname?: string;
+    } = {
+      name,
+      abbreviation: abbreviation || defaultAbbreviation,
+      SeasonTeamId,
+      UserId,
+    };
+    if (nickname !== null) {
+      whereClause.nickname = nickname;
+    }
     const [member, _created] = await Member.findOrCreate({
-      where: {
-        name,
-        nickname,
-        abbreviation: abbreviation || defaultAbbreviation,
-        SeasonTeamId,
-        UserId,
-      },
+      where: whereClause,
     });
     return member;
   }
@@ -142,12 +167,17 @@ class MemberService {
    * @param {boolean} options.all - Whether to update all members or only those associated with the user.
    * @returns {Promise<Object>} The updated member object.
    */
-  async update(id: string, data: object, UserId: string | null, options = { all: false }) {
+  async update(
+    id: string,
+    data: object,
+    UserId: string | null,
+    options = { all: false },
+  ) {
     logger.debug(
-      `MemberService update ${JSON.stringify({ id, data, UserId })}`
+      `MemberService update ${JSON.stringify({ id, data, UserId })}`,
     );
     return Member.findOne({
-      where: options.all || !UserId ? { id } : { id, UserId }
+      where: options.all || !UserId ? { id } : { id, UserId },
     }) // njsscan-ignore: node_nosqli_injection
       .then(async (foundMember) => {
         if (foundMember) {
@@ -156,17 +186,13 @@ class MemberService {
           return Member.findOne({
             where: options.all
               ? { id }
-              : (UserId !== null
+              : UserId !== null
                 ? { id, UserId }
-                : { id })
+                : { id },
           }); // njsscan-ignore: node_nosqli_injection
         } else {
-          logger.error(
-            `No member found with ID ${id} when updating`
-          );
-          throw new Error(
-            `No member found with ID ${id} when updating`
-          );
+          logger.error(`No member found with ID ${id} when updating`);
+          throw new Error(`No member found with ID ${id} when updating`);
         }
       });
   }
@@ -180,22 +206,16 @@ class MemberService {
    * @returns {Promise<void>} Resolves when the member is removed.
    */
   async remove(id: string, UserId: string | null, options = { all: false }) {
-    logger.debug(
-      `MemberService remove ${JSON.stringify({ id, UserId })}`
-    );
+    logger.debug(`MemberService remove ${JSON.stringify({ id, UserId })}`);
     return Member.findOne({
-      where: options.all || !UserId ? { id } : { id, UserId }
+      where: options.all || !UserId ? { id } : { id, UserId },
     }) // njsscan-ignore: node_nosqli_injection
       .then((foundMember) => {
         if (foundMember) {
           return foundMember.destroy();
         } else {
-          logger.error(
-            `No member found with ID ${id} when deleting`
-          );
-          throw new Error(
-            `No member found with ID ${id} when deleting`
-          );
+          logger.error(`No member found with ID ${id} when deleting`);
+          throw new Error(`No member found with ID ${id} when deleting`);
         }
       });
   }
