@@ -1,9 +1,9 @@
 import SeasonTeam from "../db/models/seasonTeam";
 import Team from "../db/models/team";
+import SeasonTeamService from "./SeasonTeamService";
 
 const { Op } = require("sequelize");
 const { logger } = require("../plugins/winston");
-const SeasonTeamService = require("./SeasonTeamService");
 
 const defaultInclude = [
   {
@@ -27,7 +27,7 @@ class TeamService {
    * @returns {Promise<Array>} List of teams.
    */
   async getAll(UserId: string | null, options = { all: false }) {
-    logger.debug(`TeamService getAll ${JSON.stringify({ UserId, options })}`)
+    logger.debug(`TeamService getAll ${JSON.stringify({ UserId, options })}`);
     return Team.findAll({
       where: options.all ? {} : UserId ? { UserId } : {},
       include: defaultInclude,
@@ -41,7 +41,7 @@ class TeamService {
    * @returns {Promise<Array>} List of teams.
    */
   async findByName(name: string, UserId: string) {
-    logger.debug(`TeamService findByName ${JSON.stringify({ name, UserId })}`)
+    logger.debug(`TeamService findByName ${JSON.stringify({ name, UserId })}`);
     return Team.findAll({
       where: { name, UserId },
       include: defaultInclude,
@@ -55,7 +55,7 @@ class TeamService {
    * @returns {Promise<Object>} The team object.
    */
   async findById(id: string, UserId: string) {
-    logger.debug(`TeamService findById ${JSON.stringify({ id, UserId })}`)
+    logger.debug(`TeamService findById ${JSON.stringify({ id, UserId })}`);
     return Team.findOne({
       where: { id, UserId },
       include: defaultInclude,
@@ -67,7 +67,7 @@ class TeamService {
    * @returns {Promise<number>} The count of teams.
    */
   getCount() {
-    logger.debug(`TeamService getCount`)
+    logger.debug(`TeamService getCount`);
     return Team.count();
   }
 
@@ -76,16 +76,20 @@ class TeamService {
    * @returns {Promise<number>} The trend value.
    */
   getTrend() {
-    logger.debug(`TeamService getTrend`)
+    logger.debug(`TeamService getTrend`);
     return Promise.all([
       Team.count({
         where: {
-          createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          createdAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
       Team.count({
         where: {
-          deletedAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          deletedAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
     ]).then(([created, deleted]) => created - deleted);
@@ -101,13 +105,18 @@ class TeamService {
    */
   async create(name: string, ClubId: string, SeasonId: string, UserId: string) {
     logger.debug(
-      `TeamService create ${JSON.stringify({ name, ClubId, SeasonId, UserId })}`
+      `TeamService create ${JSON.stringify({
+        name,
+        ClubId,
+        SeasonId,
+        UserId,
+      })}`,
     );
-    return Team.create({ name, ClubId, UserId }).then((team: Team) =>
-      SeasonTeamService.create(team.id, SeasonId, [], UserId).then(
-        (_seasonTeam: SeasonTeam) => this.findById(team.id, UserId)
-      )
-    );
+    return Team.create({ name, ClubId, UserId }).then((team: Team) => {
+      return SeasonTeamService.create(team.id, SeasonId, [], UserId).then(
+        (_seasonTeam: SeasonTeam | null) => this.findById(team.id, UserId),
+      );
+    });
   }
 
   /**
@@ -119,7 +128,7 @@ class TeamService {
    */
   async findOrCreate(name: string, ClubId: string, UserId: string) {
     logger.debug(
-      `TeamService findOrCreate ${JSON.stringify({ name, ClubId, UserId })}`
+      `TeamService findOrCreate ${JSON.stringify({ name, ClubId, UserId })}`,
     );
     const [team, _created] = await Team.findOrCreate({
       where: { name, ClubId, UserId },
@@ -136,12 +145,15 @@ class TeamService {
    * @param {boolean} options.all - Whether to update all teams.
    * @returns {Promise<Object>} The updated team object.
    */
-  async update(id: string, data: object, UserId: string | null, options = { all: false }) {
-    logger.debug(
-      `TeamService update ${JSON.stringify({ id, data, UserId })}`
-    );
+  async update(
+    id: string,
+    data: object,
+    UserId: string | null,
+    options = { all: false },
+  ) {
+    logger.debug(`TeamService update ${JSON.stringify({ id, data, UserId })}`);
     return Team.findOne({
-      where: options.all || !UserId ? { id } : { id, UserId }
+      where: options.all || !UserId ? { id } : { id, UserId },
     }) // njsscan-ignore: node_nosqli_injection
       .then(async (foundTeam: Team | null) => {
         if (foundTeam) {
@@ -163,8 +175,12 @@ class TeamService {
    * @returns {Promise<void>} Resolves when the team is removed.
    */
   async remove(id: string, UserId: string | null, options = { all: false }) {
-    logger.debug(`TeamService remove ${JSON.stringify({ id, UserId, options })}`);
-    return Team.findOne({ where: options.all || !UserId ? { id } : { id, UserId } }) // njsscan-ignore: node_nosqli_injection
+    logger.debug(
+      `TeamService remove ${JSON.stringify({ id, UserId, options })}`,
+    );
+    return Team.findOne({
+      where: options.all || !UserId ? { id } : { id, UserId },
+    }) // njsscan-ignore: node_nosqli_injection
       .then((foundTeam) => {
         if (foundTeam) {
           return foundTeam.destroy();
