@@ -19,7 +19,7 @@ class UserService {
    * @returns {Promise<Array>} Array of user objects.
    */
   async getAll() {
-    logger.debug(`UserService getAll`)
+    logger.debug(`UserService getAll`);
     return User.findAll();
   }
 
@@ -29,7 +29,7 @@ class UserService {
    * @returns {Promise<Object>} The user object.
    */
   async findById(id: string) {
-    logger.debug(`UserService findById ${JSON.stringify({ id })}`)
+    logger.debug(`UserService findById ${JSON.stringify({ id })}`);
     return User.findByPk(id, { include: ["Clubs"] });
   }
 
@@ -42,9 +42,14 @@ class UserService {
    */
   async findByUsernameOrEmail(
     usernameOrEmail: string,
-    { scope = "defaultScope" } = {}
+    { scope = "defaultScope" } = {},
   ) {
-    logger.debug(`UserService findByUsernameOrEmail ${JSON.stringify({ usernameOrEmail, scope })}`)
+    logger.debug(
+      `UserService findByUsernameOrEmail ${JSON.stringify({
+        usernameOrEmail,
+        scope,
+      })}`,
+    );
     return User.scope(scope).findOne({
       where: {
         [Op.or]: [{ username: usernameOrEmail }, { email: usernameOrEmail }],
@@ -57,7 +62,7 @@ class UserService {
    * @returns {Promise<number>} The total count of users.
    */
   getCount() {
-    logger.debug(`UserService getCount`)
+    logger.debug(`UserService getCount`);
     return User.count();
   }
 
@@ -66,16 +71,20 @@ class UserService {
    * @returns {Promise<number>} The trend value.
    */
   getTrend() {
-    logger.debug(`UserService getTrend`)
+    logger.debug(`UserService getTrend`);
     return Promise.all([
       User.count({
         where: {
-          createdAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          createdAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
       User.count({
         where: {
-          deletedAt: { [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30 },
+          deletedAt: {
+            [Op.gt]: new Date().valueOf() - 1000 * 60 * 60 * 24 * 30,
+          },
         },
       }),
     ]).then(([created, deleted]) => created - deleted);
@@ -90,19 +99,33 @@ class UserService {
    * @param {string} locale - The locale for notifications.
    * @returns {Promise<Object>} The created user object.
    */
-  async create(username: string, password: string, email: string, emailConfirmed: boolean, locale: string) {
-    logger.debug(`UserService create ${JSON.stringify({ username, password: password ? '<redacted>' : 'undefined', email, emailConfirmed, locale })}`)
+  async create(
+    username: string,
+    password: string,
+    email: string,
+    emailConfirmed: boolean,
+    locale: string,
+  ) {
+    logger.debug(
+      `UserService create ${JSON.stringify({
+        username,
+        password: password ? "<redacted>" : "undefined",
+        email,
+        emailConfirmed,
+        locale,
+      })}`,
+    );
     return User.create({ username, password, email, emailConfirmed }).then(
       (user) => {
         MailService.sendUserRegistrationNotice(
           user.username,
           user.id,
-          user.email
+          user.email,
         ).catch(logger.error);
         NotificationService.createOne(
           i18n.__({ phrase: "notifications.welcome.title", locale }),
           i18n.__({ phrase: "notifications.welcome.message", locale }),
-          user.id
+          user.id,
         );
 
         if (email) {
@@ -110,17 +133,17 @@ class UserService {
             user.username,
             user.id,
             user.email,
-            locale
+            locale,
           ).catch(logger.error);
           MailService.sendWelcomeEmail(
             user.username,
             user.id,
             user.email,
-            locale
+            locale,
           ).catch(logger.error);
         }
         return user;
-      }
+      },
     );
   }
 
@@ -132,7 +155,7 @@ class UserService {
    */
   async findOrCreate(username: string, password: string) {
     logger.debug(
-      `UserService findOrCreate ${JSON.stringify({ username, password })}`
+      `UserService findOrCreate ${JSON.stringify({ username, password })}`,
     );
     const [user, _created] = await User.findOrCreate({
       where: { username },
@@ -151,7 +174,7 @@ class UserService {
    * @returns {Promise<Object>} The updated user object.
    */
   async update(id: string, data: Record<string, unknown>) {
-    const { password, ...logData } = data
+    const { password, ...logData } = data;
     logger.debug(`UserService update ${JSON.stringify({ id, data: logData })}`);
     return User.findByPk(id).then(async (foundUser) => {
       if (foundUser) {
@@ -187,7 +210,7 @@ class UserService {
    * @returns {Promise<number>} The percentage of logged-in users.
    */
   getLoggedInPercentage() {
-    logger.debug(`UserService getLoggedInPercentage`)
+    logger.debug(`UserService getLoggedInPercentage`);
     return Promise.all([
       this.getCount(),
       User.count({
@@ -198,7 +221,7 @@ class UserService {
     ]).then(([totalCount, notLoggedInCount]) => {
       return roundToDecimals(
         ((totalCount - notLoggedInCount) / totalCount) * 100,
-        1
+        1,
       );
     });
   }
