@@ -1,7 +1,7 @@
 import logger from "../plugins/winston";
 
 function sequelizeDataTypeToHtmlInputType(sequelizeDataType: { key: string }) {
-  logger.debug(`Converting data type ${sequelizeDataType.key}`)
+  logger.debug(`Converting data type ${sequelizeDataType.key}`);
   switch (sequelizeDataType.key) {
     case "STRING":
       return "text";
@@ -14,6 +14,6 @@ function sequelizeDataTypeToHtmlInputType(sequelizeDataType: { key: string }) {
     default:
       return "text";
   }
-};
+}
 
-export default sequelizeDataTypeToHtmlInputType
+export default sequelizeDataTypeToHtmlInputType;
