@@ -7,18 +7,18 @@ const MILLIS_PER_DAY = 24 * MILLIS_PER_HOUR;
 
 export function timeStringToMillis(timeString: string) {
   let result = 0;
-  Array.from(timeString
-    .matchAll(/(\d{1,3})s/g))
-    .forEach((match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_SECOND));
-  Array.from(timeString
-    .matchAll(/(\d{1,3})m/g))
-    .forEach((match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_MINUTE));
-  Array.from(timeString
-    .matchAll(/(\d{1,3})h/g))
-    .forEach((match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_HOUR));
-  Array.from(timeString
-    .matchAll(/(\d{1,3})d/g))
-    .forEach((match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_DAY));
-  logger.debug(`Converted time ${timeString} to ${result}`)
+  Array.from(timeString.matchAll(/(\d{1,3})s/g)).forEach(
+    (match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_SECOND),
+  );
+  Array.from(timeString.matchAll(/(\d{1,3})m/g)).forEach(
+    (match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_MINUTE),
+  );
+  Array.from(timeString.matchAll(/(\d{1,3})h/g)).forEach(
+    (match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_HOUR),
+  );
+  Array.from(timeString.matchAll(/(\d{1,3})d/g)).forEach(
+    (match: string[]) => (result += parseInt(match[1]) * MILLIS_PER_DAY),
+  );
+  logger.debug(`Converted time ${timeString} to ${result}`);
   return result;
 }
